requirements.txt
# Core dependencies
pyyaml>=6.0
python-dotenv>=1.0.0

# LLM providers (optional)
google-generativeai>=0.3.0
openai>=1.0.0
anthropic>=0.7.0

# RAG dependencies
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
numpy>=1.24.0
